// src/types/impl/lib/impl/local.ts
// ExcelTables likely will need to be updated as new tables are added to the game data
// https://github.com/Kengxxiao/ArknightsGameData_YoStar/tree/main/en_US/gamedata/excel
export enum ExcelTables {
    ACTIVITY_TABLE = "activity_table",
    AUDIO_DATA = "audio_data",
    BATTLE_EQUIP_TABLE = "battle_equip_table",
    BUILDING_DATA = "building_data",
    CAMPAIGN_TABLE = "campaign_table",
    CHAPTER_TABLE = "chapter_table",
    CHAR_META_TABLE = "char_meta_table",
    CHAR_PATCH_TABLE = "char_patch_table",
    CHARACTER_TABLE = "character_table",
    CHARM_TABLE = "charm_table",
    CHARWORD_TABLE = "charword_table",
    CHECKIN_TABLE = "checkin_table",
    CLIMB_TOWER_TABLE = "climb_tower_table",
    CLUE_DATA = "clue_data",
    CRISIS_TABLE = "crisis_table",
    CRISIS_V2_TABLE = "crisis_v2_table",
    DATA_VERSION = "data_version", // .txt file
    DISPLAY_META_TABLE = "display_meta_table",
    ENEMY_HANDBOOK_TABLE = "enemy_handbook_table",
    FAVOR_TABLE = "favor_table",
    GACHA_TABLE = "gacha_table",
    GAMEDATA_CONST = "gamedata_const",
    HANDBOOK_INFO_TABLE = "handbook_info_table",
    HANDBOOK_TABLE = "handbook_table",
    HANDBOOK_TEAM_TABLE = "handbook_team_table",
    ITEM_TABLE = "item_table",
    MEDAL_TABLE = "medal_table",
    MISSION_TABLE = "mission_table",
    OPEN_SERVER_TABLE = "open_server_table",
    PLAYER_AVATAR_TABLE = "player_avatar_table",
    RANGE_TABLE = "range_table",
    REPLICATE_TABLE = "replicate_table",
    RETRO_TABLE = "retro_table",
    ROGUELIKE_TABLE = "roguelike_table",
    ROGUELIKE_TOPIC_TABLE = "roguelike_topic_table",
    SANDBOX_PERM_TABLE = "sandbox_perm_table",
    SANDBOX_TABLE = "sandbox_table",
    SHOP_CLIENT_TABLE = "shop_client_table",
    SKILL_TABLE = "skill_table",
    SKIN_TABLE = "skin_table",
    STAGE_TABLE = "stage_table",
    STORY_REVIEW_META_TABLE = "story_review_meta_table",
    STORY_REVIEW_TABLE = "story_review_table",
    STORY_TABLE = "story_table",
    TECH_BUFF_TABLE = "tech_buff_table",
    TIP_TABLE = "tip_table",
    TOKEN_TABLE = "token_table",
    UNIEQUIP_DATA = "uniequip_data",
    UNIEQUIP_TABLE = "uniequip_table",
    ZONE_TABLE = "zone_table",
}